test(task-repository): 5ファイル分割・§BUG-TR-002凍結・4テーブル確認に更新
ヘルスバーグ[却下]3点対応 (Issue #35 PR #52):

1. 500行ルール適用 + 5ファイル分割:
   - test_protocol_crud.py: 617行→445行 (TestSave9StageSemantics 削除)
   - test_save_child_tables.py: 新規作成 (TC-UT-TR-005/005b/005c 移動)
   - test_count_methods.py: 新規作成 (TC-UT-TR-006/007 移動)
   - test_find_blocked.py: find_blocked 専用に縮小
   - test_masking_fields.py: 2カラム (conversation_messages 除外)

2. §BUG-TR-002凍結適用:
   - TestConversationMessageBodyMarkdownMasked 削除
   - TestThreeColumnSimultaneousMasking 削除
   - conversations/ConversationRow imports 削除
   - MaskedText 2カラム (tasks.last_error + deliverables.body_markdown) のみ

3. TC-IT-TR-001 4テーブル確認:
   - test_alembic_task.py: 6テーブル→4テーブル期待値に変更
   - conversations/conversation_messages 非存在確認テスト追加
   - TC-IT-TR-005 (upgrade/downgrade) も4テーブルに更新

4. storage.md doc test更新:
   - test_storage_md_back_index.py: conversation_messages行を
     §BUG-TR-002 凍結確認に変更 (MaskedText→BUG-TR-002 co-location)

NOTE: test_alembic_task.py の conversations 非存在確認テスト2件は
ヴァンロッサムの 0007_task_aggregate.py migration 変更後にパス。
